docfusion/requirements.txt
scikit-learn
Pillow
pytesseract
requests
xgboost
pandas
numpy
streamlit
gradio
joblib
torch
torchvision
transformers
matplotlib
datasets
tqdm
